media-azan/requirements.txt
aiomqtt==2.2.0
python-dotenv==1.0.1
qrcode==7.4.2
telethon==1.34.0
